Added a FieldOfObservation class to display what the satellite see in orbit.
Added a FieldOfObservationSatellite tutorial to show how to display the field of observation of the satellite, modulated by the attitude.

Added an AttitudePointing class, to display the line of sight of the satellite on ground.

Added a SinusoidalAttitude tutorial, to see how to setup a sinusoidal attitude for a satellite.

Refactor of the organisation of tutorials