Added a referenceSystem.glb file to represent in 3D the reference system of a satellite
Rework of the SatelliteReferenceSystem class to display the reference system of the satellite. It will follow now the attitude of the satellite.

Rework the default resources path in the header class, wrong slash.

Added a CompensationOrbitAttitude to display a SSO orbit and display the ground track.

Modified the AbstractPointOnEarth class to use only the 10 first geodetic points for the id.

Added a constructor for CzmlModel to admit scales, maximumScales and minimumPixelSizes.

Added the displaySatelliteReferenceSystem to the SinusoidalAttitude tutorial.