REFACTOR : CzmlFileBuilder. Rethink of the way of building a CzmlFile : Now CzmlFile objects can be initiated with a Builder, and have access to 'with' functions to add objects. This way the CzmlFile is build at the end of the program and has access to the write() function.
All tutorials have been changed to fit the CzmlFileBuilder usage.

Refactor of the Satellite class : Now the class has a SatelliteBuilder class to build it. Satellites objects are now only buildable with a boundedPropagator and a finalDate (AbsoluteDate) and other arguments are called with the 'with' function. The satellite builder can now take a start date as an input

All tutorials have been rework to use only boundedPropagators or the basic constructor of the Satellite class.

The Constellation class now only has builders that accepts a list of propagators too.

Fix bug that did not display line of visibility at the end of the simulation when not entirely in the range of time of the simulation.

An AttitudePointingBuilder has been added, now AttitudePointing objects can be built with this builder, and optional arguments can be added with 'with' functions.

A CovarianceDisplayBuilder has been added, now CovarianceDisplay objects can be built with this builder, and optional arguments can be added with 'with' functions.

A FieldOfObservationBuilder has been added now FieldOfObservation objects can be built with this builder, and optional arguments can be added with 'with' functions.

A class OemAdaptor has been added, it allows the user to extract a propagator and a finalDate from an Oem file. So it can allow the build of a satellite object.

An OemAdaptorExample tutorial has been added to show how tu use the OemAdaptor to build a satellite object from an Oem file.

The TypeOfVisu enum has been removed, the line of visibility are now computed satellite by satellite, station by station.

An OemTest.xml file has been added to give an example of what an Oem file looks like.