Added a NonVectorPolylineBuilder and a VectorPolylineBuilder to build polylines separately as a vector or not with builder and 'with' functions.
Add a ManeuverSequence object, it allows the user to add maneuver Orekit objects to a propagator, use it into the satellite then display it with the ManeuverSequence object into the simulation. The basic 3D model used to represents the maneuver is now added as a maneuver_model.glb into the resources package.

A ManeuverSequenceExample tutorial has been added to show how to use a ManeuverSequence object.