Added a maneuver sequence builder to build efficiently the maneuver sequences.
The ManeuverSequence object now that an AttitudesSequence object, a lis of maneuver, a Satellite and a direction for the acceleration. The ManeuverSequenceBuilder also now accepts a boolean to display the arrow as the acceleration, or as the thrust.

The ManeuverSequence object is now fixed for the display direction of the arrow. A tutorial ManeuverSequenceExample is not available to show how to use maneuvers.

The LineOfVisibility has been fixed, it didn't display line correctly.

Fixed the ISS 3D model and the Reference System Model that didn't fit the same reference system.